+++ src/ports/SkFontMgr_win_dw_factory.cpp
@@ -11,7 +11,7 @@
 #include "include/core/SkFontMgr.h"
 #include "include/ports/SkTypeface_win.h"
 
-#if !defined(SK_DISABLE_LEGACY_FONTMGR_REFDEFAULT)
+#if !defined(SK_DISABLE_LEGACY_FONTMGR_FACTORY)
 sk_sp<SkFontMgr> SkFontMgr::Factory() {
     return SkFontMgr_New_DirectWrite();
 }