+++ modules/canvaskit/CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Passing a `null` `Typeface` to the `Font` constructor no longer uses the default typeface. See
    `CanvasKit.Typeface.GetDefault()` as a way to get the compiled-in typeface to replace this
    behavior.
+ - `MakeManagedAnimation` no longer falls back to the built-in typeface if the provided
+   FreeType data is not in the assets map.
 
 ### Added
  - `CanvasKit.Typeface.GetDefault()` as a way to explicitly get the compiled-in typeface (if any).